[ADDED] Server-wide rate limiting
This is a very basic and not necessarily accurate server-wide rate
limiting. Both ingress and egress are counted toward the limit.
The changes when running without rate limit should not have any
performance impact. When rate limiting is enabled, but limit is
very high, a performance degradation is possible (cost of 3 atomic
operations).